refactor: move transformer to storybook-helpers
